+++ R/AFM.getLine.R
@@ -16,7 +16,7 @@
 #' @seealso \code{\link{AFM.lineProfile}}
 #'
 #' @examples
-#' filename = AFM.getSampleImages()
+#' filename = AFM.getSampleImages('ibw')[1]
 #' afmd = AFM.import(filename)
 #' afmd2 = AFM.getLine(afmd, 50)
 #' plot(afmd2, addLines = TRUE)